Extract SteppedSlider component
